refactor: replace SessionStoreV2 facade with ConversationService
SessionStoreV2 is now a pure utility class (private constructor, static
methods only) used exclusively for legacy JSONL parsing during migration.

All runtime callers (11 production files + 3 test files) now use
ConversationService — a @Service(Service.Level.PROJECT) class that
provides the public API for conversation persistence via SQLite.

Changes:
- Create ConversationService in session.db package with session ID
  management, async write queue, and entry read/write operations
- Remove SessionStoreV2 from plugin.xml service registration
- Migrate AcpClient, ChatWebServer, SearchConversationHistoryTool,
  SessionSwitchService, UsageStatisticsLoader, BackfillMiner,
  PromptsPanel, HistoryContextWindow, AcpConnectPanel,
  ChatToolWindowContent, and MemorySettingsConfigurable
- Update backfill services to use static SessionStoreV2 methods directly
- Move InstanceMethodTests from SessionStoreV2Test to new
  ConversationServiceTest (22 tests, all passing)
- Update platform tests to verify ConversationService resolves
