+++ plugin-core/src/main/java/com/github/catatafishen/agentbridge/services/ToolCallStatisticsBackfill.java
@@ -63,8 +63,7 @@ public final class ToolCallStatisticsBackfill {
      */
     public static BackfillResult backfill(@NotNull ToolCallStatisticsService service,
                                           @NotNull String basePath) {
-        SessionStoreV2 store = new SessionStoreV2();
-        List<SessionStoreV2.SessionRecord> sessions = store.listSessions(basePath);
+        List<SessionStoreV2.SessionRecord> sessions = SessionStoreV2.listSessionsFromJsonlIndex(basePath);
         if (sessions.isEmpty()) {
             LOG.info("Backfill: no sessions found");
             return new BackfillResult(0, 0, 0);
