+++ plugin-core/src/main/java/com/github/catatafishen/agentbridge/session/SessionSwitchService.java
@@ -5,6 +5,7 @@ import com.github.catatafishen.agentbridge.agent.codex.CodexAppServerClient;
 import com.github.catatafishen.agentbridge.psi.PlatformApiCompat;
 import com.github.catatafishen.agentbridge.services.AgentProfileManager;
 import com.github.catatafishen.agentbridge.services.GenericSettings;
+import com.github.catatafishen.agentbridge.session.db.ConversationService;
 import com.github.catatafishen.agentbridge.session.exporters.AnthropicClientExporter;
 import com.github.catatafishen.agentbridge.session.exporters.ClaudeCliExporter;
 import com.github.catatafishen.agentbridge.session.exporters.CodexClientExporter;
@@ -12,7 +13,6 @@ import com.github.catatafishen.agentbridge.session.exporters.CopilotClientExport
 import com.github.catatafishen.agentbridge.session.exporters.ExportUtils;
 import com.github.catatafishen.agentbridge.session.exporters.KiroClientExporter;
 import com.github.catatafishen.agentbridge.session.exporters.OpenCodeClientExporter;
-import com.github.catatafishen.agentbridge.session.v2.SessionStoreV2;
 import com.github.catatafishen.agentbridge.settings.JunieClientConfigurable;
 import com.github.catatafishen.agentbridge.settings.KiroClientConfigurable;
 import com.github.catatafishen.agentbridge.settings.OpenCodeClientConfigurable;
@@ -180,7 +180,7 @@ public final class SessionSwitchService implements Disposable {
         // Wait for any in-flight saveAsync to flush the v2 JSONL to disk before we read it.
         // Without this, the export may read stale data if the user switches agents immediately
         // after a conversation turn completes.
-        SessionStoreV2.getInstance(project).awaitPendingSave(5_000);
+        ConversationService.getInstance(project).awaitPendingSave(5_000);
 
         // Load v2 session — this is our source of truth, kept up-to-date by the plugin
         // on every conversation save. No need to re-import from the previous client's
@@ -620,8 +620,8 @@ public final class SessionSwitchService implements Disposable {
 
     @NotNull
     private List<EntryData> loadCurrentV2Session(@Nullable String basePath) {
-        SessionStoreV2.RecentEntriesResult result =
-            SessionStoreV2.getInstance(project).loadRecentEntries(basePath);
+        ConversationService.RecentEntriesResult result =
+            ConversationService.getInstance(project).loadRecentEntries(basePath);
         return result != null ? result.entries() : List.of();
     }
