+++ plugin-core/src/main/java/com/github/catatafishen/agentbridge/ui/statistics/UsageStatisticsLoader.java
@@ -2,10 +2,10 @@ package com.github.catatafishen.agentbridge.ui.statistics;
 
 import com.github.catatafishen.agentbridge.services.AgentIdMapper;
 import com.github.catatafishen.agentbridge.services.ToolCallStatisticsService;
+import com.github.catatafishen.agentbridge.session.db.ConversationService;
 import com.github.catatafishen.agentbridge.session.exporters.ExportUtils;
 import com.github.catatafishen.agentbridge.session.v2.EntryDataJsonAdapter;
 import com.github.catatafishen.agentbridge.session.v2.SessionFileRotation;
-import com.github.catatafishen.agentbridge.session.v2.SessionStoreV2;
 import com.github.catatafishen.agentbridge.ui.EntryData;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
@@ -126,9 +126,9 @@ final class UsageStatisticsLoader {
         Map<String, String> agentDisplayNames = new LinkedHashMap<>();
         Set<String> agentIds = new LinkedHashSet<>();
 
-        List<SessionStoreV2.SessionRecord> sessions =
-            SessionStoreV2.getInstance(project).listSessions(basePath);
-        for (SessionStoreV2.SessionRecord session : sessions) {
+        List<ConversationService.SessionRecord> sessions =
+            ConversationService.getInstance(project).listSessions();
+        for (ConversationService.SessionRecord session : sessions) {
             String agentId = toAgentId(session.agent());
             agentIds.add(agentId);
             agentDisplayNames.putIfAbsent(agentId, session.agent());
@@ -175,8 +175,8 @@ final class UsageStatisticsLoader {
         @NotNull LocalDate endDate) {
 
         String basePath = project.getBasePath();
-        List<SessionStoreV2.SessionRecord> sessions =
-            SessionStoreV2.getInstance(project).listSessions(basePath);
+        List<ConversationService.SessionRecord> sessions =
+            ConversationService.getInstance(project).listSessions();
         if (sessions.isEmpty()) {
             LOG.info("Statistics: no sessions found for basePath=" + basePath);
             return emptySnapshot(startDate, endDate);
@@ -188,7 +188,7 @@ final class UsageStatisticsLoader {
 
         File sessionsDir = ExportUtils.sessionsDir(basePath);
 
-        for (SessionStoreV2.SessionRecord session : sessions) {
+        for (ConversationService.SessionRecord session : sessions) {
             String agentDisplay = session.agent();
             String fallbackAgentId = toAgentId(agentDisplay);
             agentIds.add(fallbackAgentId);
