+++ plugin-core/src/test/java/com/github/catatafishen/agentbridge/session/v2/SessionStoreV2Test.java
@@ -1,6 +1,5 @@
 package com.github.catatafishen.agentbridge.session.v2;
 
-import com.github.catatafishen.agentbridge.session.db.ConversationDatabase;
 import com.github.catatafishen.agentbridge.ui.ContextFileRef;
 import com.github.catatafishen.agentbridge.ui.EntryData;
 import com.github.catatafishen.agentbridge.ui.FileRef;
@@ -8,26 +7,14 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.sql.Connection;
-import java.sql.DriverManager;
 import java.util.List;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -1252,380 +1239,5 @@ class SessionStoreV2Test {
         msg.add("parts", parts);
         return msg;
     }
-
-    // ══════════════════════════════════════════════════════════════════════════
-    // Instance method tests — SQLite-backed via in-memory database
-    // ══════════════════════════════════════════════════════════════════════════
-
-    @Nested
-    @DisplayName("InstanceMethodTests")
-    class InstanceMethodTests {
-
-        @TempDir
-        Path tempDir;
-
-        private ConversationDatabase database;
-
-        @BeforeEach
-        void setUp() throws Exception {
-            database = new ConversationDatabase();
-            Connection conn = DriverManager.getConnection("jdbc:sqlite::memory:");
-            database.initializeWithConnection(conn);
-        }
-
-        private SessionStoreV2 newStore() {
-            SessionStoreV2 store = new SessionStoreV2(database);
-            store.setCurrentAgent("test-agent");
-            return store;
-        }
-
-        private Path currentSessionIdFile() {
-            return tempDir.resolve(".agent-work").resolve("sessions").resolve(".current-session-id");
-        }
-
-        // ── getCurrentSessionId ───────────────────────────────────────────────
-
-        @Test
-        @DisplayName("getCurrentSessionId creates the .current-session-id file on first call")
-        void getCurrentSessionId_createsFileOnFirstCall() {
-            SessionStoreV2 store = newStore();
-
-            String id = store.getCurrentSessionId(tempDir.toString());
-
-            assertNotNull(id);
-            assertFalse(id.isBlank());
-            assertTrue(currentSessionIdFile().toFile().exists(),
-                ".current-session-id file should exist after first call");
-        }
-
-        @Test
-        @DisplayName("getCurrentSessionId returns the same ID on repeated calls (cached on disk)")
-        void getCurrentSessionId_returnsSameIdOnRepeatedCalls() {
-            SessionStoreV2 store = newStore();
-
-            String first = store.getCurrentSessionId(tempDir.toString());
-            String second = store.getCurrentSessionId(tempDir.toString());
-            String third = store.getCurrentSessionId(tempDir.toString());
-
-            assertEquals(first, second);
-            assertEquals(second, third);
-        }
-
-        @Test
-        @DisplayName("after resetCurrentSessionId the next call generates a new ID")
-        void getCurrentSessionId_newIdAfterReset() {
-            SessionStoreV2 store = newStore();
-
-            String original = store.getCurrentSessionId(tempDir.toString());
-            store.resetCurrentSessionId(tempDir.toString());
-            String fresh = store.getCurrentSessionId(tempDir.toString());
-
-            assertNotEquals(original, fresh, "a new UUID should be generated after reset");
-        }
-
-        @Test
-        @DisplayName("two different basePaths produce different session IDs")
-        void getCurrentSessionId_differentBasePaths() throws IOException {
-            SessionStoreV2 store = newStore();
-            Path otherDir = Files.createTempDirectory("other");
-
-            String id1 = store.getCurrentSessionId(tempDir.toString());
-            String id2 = store.getCurrentSessionId(otherDir.toString());
-
-            assertNotEquals(id1, id2, "different base paths should produce different session IDs");
-        }
-
-        // ── resetCurrentSessionId ─────────────────────────────────────────────
-
-        @Test
-        @DisplayName("resetCurrentSessionId deletes the .current-session-id file when it exists")
-        void resetCurrentSessionId_deletesFile() {
-            SessionStoreV2 store = newStore();
-            store.getCurrentSessionId(tempDir.toString()); // create the file
-            assertTrue(currentSessionIdFile().toFile().exists(), "prerequisite: file must exist");
-
-            store.resetCurrentSessionId(tempDir.toString());
-
-            assertFalse(currentSessionIdFile().toFile().exists(), "file should be deleted");
-        }
-
-        @Test
-        @DisplayName("resetCurrentSessionId is a no-op when the file does not exist")
-        void resetCurrentSessionId_noOpWhenFileMissing() {
-            SessionStoreV2 store = newStore();
-            assertFalse(currentSessionIdFile().toFile().exists(), "prerequisite: file must not exist");
-
-            assertDoesNotThrow(() -> store.resetCurrentSessionId(tempDir.toString()));
-        }
-
-        // ── listSessions ──────────────────────────────────────────────────────
-
-        @Test
-        @DisplayName("listSessions returns empty list when no sessions exist in the database")
-        void listSessions_emptyWhenNoSessions() {
-            SessionStoreV2 store = newStore();
-
-            List<SessionStoreV2.SessionRecord> sessions = store.listSessions(tempDir.toString());
-
-            assertNotNull(sessions);
-            assertTrue(sessions.isEmpty());
-        }
-
-        @Test
-        @DisplayName("listSessions returns sessions after appendEntries creates them")
-        void listSessions_returnsSessionsAfterAppend() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Hello", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-
-            List<SessionStoreV2.SessionRecord> sessions = store.listSessions(tempDir.toString());
-
-            assertEquals(1, sessions.size());
-            assertEquals("test-agent", sessions.get(0).agent());
-            assertEquals(1, sessions.get(0).turnCount());
-        }
-
-        @Test
-        @DisplayName("listSessions returns session with display name from first prompt")
-        void listSessions_displaysNameFromFirstPrompt() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Fix the auth bug", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-
-            List<SessionStoreV2.SessionRecord> sessions = store.listSessions(tempDir.toString());
-
-            assertEquals(1, sessions.size());
-            // The display_name in SQLite comes from the first prompt text via ConversationWriter
-            assertFalse(sessions.get(0).name().isEmpty());
-        }
-
-        // ── appendEntries + loadEntries round-trip ────────────────────────────
-
-        @Test
-        @DisplayName("appendEntries + loadEntries round-trips a Prompt entry")
-        void appendAndLoad_promptRoundTrip() {
-            SessionStoreV2 store = newStore();
-            EntryData.Prompt prompt = new EntryData.Prompt("Hello world", "2024-01-01T00:00:00Z",
-                null, "eid-1", "eid-1");
-
-            store.appendEntries(tempDir.toString(), List.of(prompt));
-            List<EntryData> loaded = store.loadEntries(tempDir.toString());
-
-            assertNotNull(loaded);
-            assertFalse(loaded.isEmpty());
-            // First entry should be the prompt
-            assertTrue(loaded.stream().anyMatch(e -> e instanceof EntryData.Prompt));
-            EntryData.Prompt loadedPrompt = loaded.stream()
-                .filter(e -> e instanceof EntryData.Prompt)
-                .map(e -> (EntryData.Prompt) e)
-                .findFirst().orElseThrow();
-            assertEquals("Hello world", loadedPrompt.getText());
-        }
-
-        @Test
-        @DisplayName("appendEntries + loadEntries round-trips Text, ToolCall, and Thinking entries")
-        void appendAndLoad_multipleEntryTypesRoundTrip() {
-            SessionStoreV2 store = newStore();
-            List<EntryData> entries = List.of(
-                new EntryData.Prompt("Question", "2024-01-01T00:00:00Z", null, "p1", "p1"),
-                new EntryData.Text("Answer", "2024-01-01T00:00:01Z", "test-agent", "gpt-4", "t1"),
-                new EntryData.Thinking("hmm...", "2024-01-01T00:00:02Z", "test-agent", "gpt-4", "th1"),
-                new EntryData.ToolCall("readFile", "{\"path\":\"a.txt\"}", "filesystem",
-                    "contents", "completed", "Read a file", "/a.txt",
-                    false, null, null, "2024-01-01T00:00:03Z", "test-agent", "gpt-4", "tc1")
-            );
-
-            store.appendEntries(tempDir.toString(), entries);
-            List<EntryData> loaded = store.loadEntries(tempDir.toString());
-
-            assertNotNull(loaded);
-            // Should contain prompt + text + thinking + toolcall (maybe also TurnStats)
-            assertTrue(loaded.stream().anyMatch(e -> e instanceof EntryData.Prompt));
-            assertTrue(loaded.stream().anyMatch(e -> e instanceof EntryData.Text));
-            assertTrue(loaded.stream().anyMatch(e -> e instanceof EntryData.Thinking));
-            assertTrue(loaded.stream().anyMatch(e -> e instanceof EntryData.ToolCall));
-            EntryData.ToolCall loadedTc = loaded.stream()
-                .filter(e -> e instanceof EntryData.ToolCall)
-                .map(e -> (EntryData.ToolCall) e)
-                .findFirst().orElseThrow();
-            assertEquals("readFile", loadedTc.getTitle());
-        }
-
-        @Test
-        @DisplayName("loadEntries returns null when no session exists")
-        void loadEntries_nullWhenNoSession() {
-            SessionStoreV2 store = newStore();
-
-            List<EntryData> loaded = store.loadEntries(tempDir.toString());
-
-            // No data written yet → should be null or empty
-            // getCurrentSessionId creates a session ID but no DB rows
-            assertNull(loaded);
-        }
-
-        @Test
-        @DisplayName("multiple appendEntries calls accumulate entries")
-        void appendEntries_multipleAppendsAccumulate() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("First", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Second", "2024-01-01T00:00:01Z", null, "p2", "p2")));
-
-            List<EntryData> loaded = store.loadEntries(tempDir.toString());
-
-            assertNotNull(loaded);
-            long promptCount = loaded.stream().filter(e -> e instanceof EntryData.Prompt).count();
-            assertEquals(2, promptCount, "both prompts should be stored");
-        }
-
-        // ── sessions metadata via appendEntries ───────────────────────────────
-
-        @Test
-        @DisplayName("after appendEntries the session appears in listSessions")
-        void appendEntries_sessionAppearsInListSessions() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Fix the bug", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-
-            List<SessionStoreV2.SessionRecord> sessions = store.listSessions(tempDir.toString());
-
-            assertEquals(1, sessions.size());
-            assertEquals("test-agent", sessions.get(0).agent());
-            assertEquals(1, sessions.get(0).turnCount());
-        }
-
-        @Test
-        @DisplayName("appendEntries increments turnCount for each Prompt in the batch")
-        void appendEntries_incrementsTurnCount() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(), List.of(
-                new EntryData.Prompt("First turn", "2024-01-01T00:00:00Z", null, "p1", "p1"),
-                new EntryData.Text("Reply", "2024-01-01T00:00:01Z", "test-agent", "gpt-4", "t1"),
-                new EntryData.Prompt("Second turn", "2024-01-01T00:00:02Z", null, "p2", "p2")
-            ));
-
-            List<SessionStoreV2.SessionRecord> sessions = store.listSessions(tempDir.toString());
-            assertEquals(1, sessions.size());
-            assertEquals(2, sessions.get(0).turnCount());
-        }
-
-        @Test
-        @DisplayName("appendEntries does not overwrite existing session name on second call")
-        void appendEntries_doesNotOverwriteExistingName() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("First prompt", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Second prompt", "2024-01-01T00:00:01Z", null, "p2", "p2")));
-
-            List<SessionStoreV2.SessionRecord> sessions = store.listSessions(tempDir.toString());
-            assertFalse(sessions.isEmpty());
-            // The session name should be set from the first prompt
-            String name = sessions.get(0).name();
-            assertTrue(name.contains("First") || !name.contains("Second"),
-                "session name should be from the first prompt, not overwritten");
-        }
-
-        // ── branchCurrentSession ──────────────────────────────────────────────
-
-        @Test
-        @DisplayName("branchCurrentSession is a no-op when no current-session-id file exists")
-        void branchCurrentSession_noOpWhenNoIdFile() {
-            SessionStoreV2 store = newStore();
-            assertFalse(currentSessionIdFile().toFile().exists(), "prerequisite: no id file");
-
-            assertDoesNotThrow(() -> store.branchCurrentSession(tempDir.toString()));
-        }
-
-        @Test
-        @DisplayName("branchCurrentSession does not change the current-session-id file")
-        void branchCurrentSession_originalIdFileUnchanged() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Seed", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-            String originalId = store.getCurrentSessionId(tempDir.toString());
-
-            store.branchCurrentSession(tempDir.toString());
-
-            String idAfterBranch = store.getCurrentSessionId(tempDir.toString());
-            assertEquals(originalId, idAfterBranch, "current-session-id must not change after branching");
-        }
-
-        // ── loadEntriesBySessionId ────────────────────────────────────────────
-
-        @Test
-        @DisplayName("loadEntriesBySessionId returns entries for the given session")
-        void loadEntriesBySessionId_returnsEntriesForSession() {
-            SessionStoreV2 store = newStore();
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Hello", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-            String sessionId = store.getCurrentSessionId(tempDir.toString());
-
-            List<EntryData> entries = store.loadEntriesBySessionId(tempDir.toString(), sessionId);
-            assertNotNull(entries);
-            assertFalse(entries.isEmpty());
-            assertTrue(entries.stream().anyMatch(e -> e instanceof EntryData.Prompt));
-        }
-
-        @Test
-        @DisplayName("loadEntriesBySessionId returns null for unknown session")
-        void loadEntriesBySessionId_nullForUnknownSession() {
-            SessionStoreV2 store = newStore();
-
-            List<EntryData> entries = store.loadEntriesBySessionId(tempDir.toString(), "nonexistent-id");
-            assertNull(entries);
-        }
-
-        // ── agent preservation ────────────────────────────────────────────────
-
-        @Test
-        @DisplayName("agent is not overwritten on subsequent appendEntries calls")
-        void appendEntries_agentPreservedOnSubsequentAppend() {
-            SessionStoreV2 store = newStore();
-            store.setCurrentAgent("agent-one");
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("First prompt", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-            String sessionId = store.getCurrentSessionId(tempDir.toString());
-
-            // Switch to a different agent and append more entries.
-            store.setCurrentAgent("agent-two");
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Second prompt", "2024-01-01T00:01:00Z", null, "p2", "p2")));
-
-            Optional<SessionStoreV2.SessionRecord> rec = store.listSessions(tempDir.toString())
-                .stream().filter(r -> r.id().equals(sessionId)).findFirst();
-            assertTrue(rec.isPresent(), "session should exist");
-            assertEquals("agent-one", rec.get().agent(),
-                "original agent must not be overwritten by subsequent appends");
-        }
-
-        @Test
-        @DisplayName("appendEntries preserves original agent after agent switch")
-        void appendEntries_preservesOriginalAgent() {
-            SessionStoreV2 store = newStore(); // currentAgent = "test-agent"
-
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Prompt("Hello", "2024-01-01T00:00:00Z", null, "p1", "p1")));
-
-            // Simulate switching to a different agent mid-session
-            store.setCurrentAgent("different-agent");
-            store.appendEntries(tempDir.toString(),
-                List.of(new EntryData.Text("Reply", "2024-01-01T00:00:01Z", "different-agent", "gpt-4", "t1")));
-
-            List<SessionStoreV2.SessionRecord> sessions = store.listSessions(tempDir.toString());
-            assertEquals(1, sessions.size());
-            assertEquals("test-agent", sessions.get(0).agent(),
-                "session agent should remain from the first append, not be overwritten");
-        }
-    }
 }
+
